Refactor and document the Affordability API project.
This commit addresses a comprehensive set of cleanup, refactoring, and documentation tasks to improve project maintainability and developer experience.

Key changes include:

1.  **Frontend Cleanup:**
    -   Removed unused boilerplate files (`logo.svg`, `App.test.js`, etc.).
    -   Cleaned default styles from `App.css`.
    -   Reviewed all React components for clarity and correctness.

2.  **Backend Refactoring:**
    -   Consolidated Lambda functions into a dedicated `handlers/` directory, with each function in its own file.
    -   Created a reusable `db.js` utility for PostgreSQL connections to centralize database logic.
    -   Updated `serverless.yaml` to point to the new handler locations.
    -   Deleted the monolithic `handler.js` file.

3.  **Documentation:**
    -   Created a new `docs/` directory for all project documentation.
    -   Added `ADR-001_Tech_Stack.md` to record architectural decisions.
    -   Added `Database_Schema.md` to describe the database structure.
    -   Added `API_Endpoints.md` to document the API.

4.  **README Update:**
    -   Added a new "Project Documentation" section to the root `README.md` to make the new documentation easily discoverable.
